[ADD] event_question: extracted from website_event
