+++ addons/event_question/data/event_demo.xml
@@ -1,0 +1,82 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+    <!-- Event Type -->
+    <record id="event.event_type_0" model="event.type">
+        <field name="question_ids" eval="[(5, 0, 0),
+            (0, 0, {'title': 'Name', 'question_type': 'name', 'is_mandatory_answer': True}),
+            (0, 0, {'title': 'Email', 'question_type': 'email', 'is_mandatory_answer': True}),
+            (0, 0, {'title': 'Phone', 'question_type': 'phone'})]"/>
+    </record>
+
+    <record id="event.event_type_1" model="event.type">
+        <field name="question_ids" eval="[(5, 0, 0),
+            (0, 0, {'title': 'Name', 'question_type': 'name', 'is_mandatory_answer': True}),
+            (0, 0, {'title': 'Email', 'question_type': 'email', 'is_mandatory_answer': True}),
+            (0, 0, {'title': 'Phone', 'question_type': 'phone'})]"/>
+    </record>
+
+    <record id="event.event_type_2" model="event.type">
+        <field name="question_ids" eval="[(5, 0, 0),
+            (0, 0, {'title': 'Name', 'question_type': 'name', 'is_mandatory_answer': True}),
+            (0, 0, {'title': 'Email', 'question_type': 'email', 'is_mandatory_answer': True}),
+            (0, 0, {'title': 'Phone', 'question_type': 'phone'})]"/>
+    </record>
+
+    <!-- Event Event -->
+    <record id="event.event_0" model="event.event">
+        <field name="question_ids" eval="[(5, 0, 0),
+            (0, 0, {'title': 'Name', 'question_type': 'name', 'is_mandatory_answer': True}),
+            (0, 0, {'title': 'Email', 'question_type': 'email', 'is_mandatory_answer': True}),
+            (0, 0, {'title': 'Phone', 'question_type': 'phone'})]"/>
+    </record>
+
+    <record id="event.event_1" model="event.event">
+        <field name="question_ids" eval="[(5, 0, 0),
+            (0, 0, {'title': 'Name', 'question_type': 'name', 'is_mandatory_answer': True}),
+            (0, 0, {'title': 'Email', 'question_type': 'email', 'is_mandatory_answer': True}),
+            (0, 0, {'title': 'Phone', 'question_type': 'phone'})]"/>
+    </record>
+
+    <record id="event.event_2" model="event.event">
+        <field name="question_ids" eval="[(5, 0, 0),
+            (0, 0, {'title': 'Name', 'question_type': 'name', 'is_mandatory_answer': True}),
+            (0, 0, {'title': 'Email', 'question_type': 'email', 'is_mandatory_answer': True}),
+            (0, 0, {'title': 'Phone', 'question_type': 'phone'})]"/>
+    </record>
+
+    <record id="event.event_3" model="event.event">
+        <field name="question_ids" eval="[(5, 0, 0),
+            (0, 0, {'title': 'Name', 'question_type': 'name', 'is_mandatory_answer': True}),
+            (0, 0, {'title': 'Email', 'question_type': 'email', 'is_mandatory_answer': True}),
+            (0, 0, {'title': 'Phone', 'question_type': 'phone'})]"/>
+    </record>
+
+    <record id="event.event_4" model="event.event">
+        <field name="question_ids" eval="[(5, 0, 0),
+            (0, 0, {'title': 'Name', 'question_type': 'name', 'is_mandatory_answer': True}),
+            (0, 0, {'title': 'Email', 'question_type': 'email', 'is_mandatory_answer': True}),
+            (0, 0, {'title': 'Phone', 'question_type': 'phone'})]"/>
+    </record>
+
+    <record id="event.event_5" model="event.event">
+        <field name="question_ids" eval="[(5, 0, 0),
+            (0, 0, {'title': 'Name', 'question_type': 'name', 'is_mandatory_answer': True}),
+            (0, 0, {'title': 'Email', 'question_type': 'email', 'is_mandatory_answer': True}),
+            (0, 0, {'title': 'Phone', 'question_type': 'phone'})]"/>
+    </record>
+
+    <record id="event.event_6" model="event.event">
+        <field name="question_ids" eval="[(5, 0, 0),
+            (0, 0, {'title': 'Name', 'question_type': 'name', 'is_mandatory_answer': True}),
+            (0, 0, {'title': 'Email', 'question_type': 'email', 'is_mandatory_answer': True}),
+            (0, 0, {'title': 'Phone', 'question_type': 'phone'})]"/>
+    </record>
+
+    <record id="event.event_7" model="event.event">
+        <field name="question_ids" eval="[(5, 0, 0),
+            (0, 0, {'title': 'Name', 'question_type': 'name', 'is_mandatory_answer': True}),
+            (0, 0, {'title': 'Email', 'question_type': 'email', 'is_mandatory_answer': True}),
+            (0, 0, {'title': 'Phone', 'question_type': 'phone'})]"/>
+    </record>
+</odoo>
