+++ addons/event_question/models/__init__.py
@@ -1,0 +1,6 @@
+from . import event_question_answer
+from . import event_registration_answer
+from . import event_registration
+from . import event_type
+from . import event_event
+from . import event_question
