+++ addons/event_question/security/ir.model.access.csv
@@ -1,0 +1,11 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+access_event_question_public,event.question,model_event_question,base.group_public,1,0,0,0
+access_event_question_portal,event.question,model_event_question,base.group_portal,1,0,0,0
+access_event_question_employee,event.question,model_event_question,base.group_user,1,0,0,0
+access_event_question_user,event.question.user,model_event_question,event.group_event_user,1,1,1,1
+access_event_question_answer_public,event.question.answer,model_event_question_answer,base.group_public,1,0,0,0
+access_event_question_answer_portal,event.question.answer,model_event_question_answer,base.group_portal,1,0,0,0
+access_event_question_answer_employee,event.question.answer,model_event_question_answer,base.group_user,1,0,0,0
+access_event_question_answer_registration,event.question.answer.registration,model_event_question_answer,event.group_event_registration_desk,1,1,0,0
+access_event_question_answer_user,event.question.answer.user,model_event_question_answer,event.group_event_user,1,1,1,1
+access_event_registration_answer,event.registration.answer,model_event_registration_answer,event.group_event_registration_desk,1,1,1,1
