+++ addons/website_event/data/event_demo.xml
@@ -1,28 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
-
-    <!-- Event Type -->
-    <record id="event.event_type_0" model="event.type">
-        <field name="question_ids" eval="[(5, 0, 0),
-            (0, 0, {'title': 'Name', 'question_type': 'name', 'is_mandatory_answer': True}),
-            (0, 0, {'title': 'Email', 'question_type': 'email', 'is_mandatory_answer': True}),
-            (0, 0, {'title': 'Phone', 'question_type': 'phone'})]"/>
-    </record>
-
-    <record id="event.event_type_1" model="event.type">
-        <field name="question_ids" eval="[(5, 0, 0),
-            (0, 0, {'title': 'Name', 'question_type': 'name', 'is_mandatory_answer': True}),
-            (0, 0, {'title': 'Email', 'question_type': 'email', 'is_mandatory_answer': True}),
-            (0, 0, {'title': 'Phone', 'question_type': 'phone'})]"/>
-    </record>
-
-    <record id="event.event_type_2" model="event.type">
-        <field name="question_ids" eval="[(5, 0, 0),
-            (0, 0, {'title': 'Name', 'question_type': 'name', 'is_mandatory_answer': True}),
-            (0, 0, {'title': 'Email', 'question_type': 'email', 'is_mandatory_answer': True}),
-            (0, 0, {'title': 'Phone', 'question_type': 'phone'})]"/>
-    </record>
-
     <!-- Event Event -->
     <record id="event.event_0" model="event.event">
         <field name="website_published" eval="True"/>
@@ -60,10 +37,6 @@
                 </div>
             </div>
         </field>
-        <field name="question_ids" eval="[(5, 0, 0),
-            (0, 0, {'title': 'Name', 'question_type': 'name', 'is_mandatory_answer': True}),
-            (0, 0, {'title': 'Email', 'question_type': 'email', 'is_mandatory_answer': True}),
-            (0, 0, {'title': 'Phone', 'question_type': 'phone'})]"/>
     </record>
 
     <record id="event.event_1" model="event.event">
@@ -84,10 +57,6 @@
                 </div>
             </div>
         </field>
-        <field name="question_ids" eval="[(5, 0, 0),
-            (0, 0, {'title': 'Name', 'question_type': 'name', 'is_mandatory_answer': True}),
-            (0, 0, {'title': 'Email', 'question_type': 'email', 'is_mandatory_answer': True}),
-            (0, 0, {'title': 'Phone', 'question_type': 'phone'})]"/>
     </record>
 
     <record id="event.event_2" model="event.event">
@@ -130,10 +99,6 @@
     </div>
 </div>
         </field>
-        <field name="question_ids" eval="[(5, 0, 0),
-            (0, 0, {'title': 'Name', 'question_type': 'name', 'is_mandatory_answer': True}),
-            (0, 0, {'title': 'Email', 'question_type': 'email', 'is_mandatory_answer': True}),
-            (0, 0, {'title': 'Phone', 'question_type': 'phone'})]"/>
     </record>
 
     <record id="event.event_3" model="event.event">
@@ -152,20 +117,12 @@
                 </div>
             </div>
         </field>
-        <field name="question_ids" eval="[(5, 0, 0),
-            (0, 0, {'title': 'Name', 'question_type': 'name', 'is_mandatory_answer': True}),
-            (0, 0, {'title': 'Email', 'question_type': 'email', 'is_mandatory_answer': True}),
-            (0, 0, {'title': 'Phone', 'question_type': 'phone'})]"/>
     </record>
 
     <record id="event.event_4" model="event.event">
         <field name="website_published" eval="True"/>
         <field name="subtitle">Discover how to grow a sustainable business with our experts.</field>
         <field name="cover_properties">{"background-image": "url('/website_event/static/src/img/event_cover_4.jpg')", "resize_class": "o_record_has_cover o_half_screen_height", "opacity": "0.4"}</field>
-        <field name="question_ids" eval="[(5, 0, 0),
-            (0, 0, {'title': 'Name', 'question_type': 'name', 'is_mandatory_answer': True}),
-            (0, 0, {'title': 'Email', 'question_type': 'email', 'is_mandatory_answer': True}),
-            (0, 0, {'title': 'Phone', 'question_type': 'phone'})]"/>
     </record>
 
     <record id="event.event_5" model="event.event">
@@ -185,19 +142,11 @@
                 </div>
             </div>
         </field>
-        <field name="question_ids" eval="[(5, 0, 0),
-            (0, 0, {'title': 'Name', 'question_type': 'name', 'is_mandatory_answer': True}),
-            (0, 0, {'title': 'Email', 'question_type': 'email', 'is_mandatory_answer': True}),
-            (0, 0, {'title': 'Phone', 'question_type': 'phone'})]"/>
     </record>
 
     <record id="event.event_6" model="event.event">
         <field name="website_published" eval="False"/>
         <field name="cover_properties">{"background-image": "none", "background-color": "secondary", "opacity": ""}</field>
-        <field name="question_ids" eval="[(5, 0, 0),
-            (0, 0, {'title': 'Name', 'question_type': 'name', 'is_mandatory_answer': True}),
-            (0, 0, {'title': 'Email', 'question_type': 'email', 'is_mandatory_answer': True}),
-            (0, 0, {'title': 'Phone', 'question_type': 'phone'})]"/>
     </record>
 
     <record id="event.event_7" model="event.event">
@@ -219,10 +168,6 @@
     </div>
 </div>
         </field>
-        <field name="question_ids" eval="[(5, 0, 0),
-            (0, 0, {'title': 'Name', 'question_type': 'name', 'is_mandatory_answer': True}),
-            (0, 0, {'title': 'Email', 'question_type': 'email', 'is_mandatory_answer': True}),
-            (0, 0, {'title': 'Phone', 'question_type': 'phone'})]"/>
     </record>
 
 </odoo>
